# docs/openapi.yaml
openapi: 3.1.0
info:
  title: MDM Platform API
  description: REST API for the MDM admin console. Manages devices, policies, apps, compliance, and more.
  version: 1.0.0
  contact:
    name: MDM Platform Team
  license:
    name: Proprietary
servers:
  - url: http://localhost:3001/api/v1
    description: Local development
  - url: https://api.staging.mdm.example.com/api/v1
    description: Staging
  - url: https://api.mdm.example.com/api/v1
    description: Production
security:
  - bearerAuth: []
  - apiKeyAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
...
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
    Pagination:
      type: object
      required: [page, limit, total, totalPages]
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
    ApiError:
      type: object
      required: [statusCode, error, message]
      properties:
        statusCode:
          type: integer
        error:
          type: string
        message:
          type: string
        details:
          type: object
        requestId:
          type: string
    Platform:
      type: string
      enum: [ios, android, windows, macos, linux]
...
    AppType:
      type: string
      enum: [public, enterprise, web_clip]

    # --- Auth ---
...
    App:
      type: object
      properties:
        id:
          type: string
          format: uuid
        orgId:
          type: string
          format: uuid
        name:
          type: string
        bundleId:
          type: string
        platform:
          $ref: '#/components/schemas/Platform'
        type:
          $ref: '#/components/schemas/AppType'
        iconUrl:
          type: string
          format: uri
        description:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateAppRequest:
      type: object
      required: [name, bundleId, platform]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        bundleId:
          type: string
          minLength: 1
          maxLength: 255
        platform:
          $ref: '#/components/schemas/Platform'
        type:
          $ref: '#/components/schemas/AppType'
        iconUrl:
          type: string
          format: uri
        description:
          type: string
    UpdateAppRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        iconUrl:
          type: string
          format: uri
        description:
          type: string
    CreateAppVersionRequest:
      type: object
      required: [version]
      properties:
        version:
          type: string
          minLength: 1
        downloadUrl:
          type: string
          format: uri
        releaseNotes:
          type: string
    AppAssignmentRequest:
      type: object
      properties:
        deviceId:
          type: string
          format: uuid
        groupId:
          type: string
          format: uuid
        isRequired:
          type: boolean
          default: false
    AppVersion:
      type: object
      properties:
        id:
          type: string
          format: uuid
        appId:
          type: string
          format: uuid
        version:
          type: string
        downloadUrl:
          type: string
        releaseNotes:
          type: string
        createdAt:
          type: string
          format: date-time
    AppListResponse:
      type: object
      required: [data, pagination]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/App'
        pagination:
          $ref: '#/components/schemas/Pagination'

    # --- Kiosk Profile ---
...
    AppAssignment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        appId:
          type: string
          format: uuid
        deviceId:
          type: string
          format: uuid
        groupId:
          type: string
          format: uuid
        isRequired:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    NoContent:
      description: Successful deletion, no content returned
paths:
...
  /devices/{id}/apps:
    get:
      operationId: getDeviceApps
      summary: Get apps assigned to a device
      tags: [devices]
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Assigned apps
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        assignmentId:
                          type: string
                          format: uuid
                        appId:
                          type: string
                          format: uuid
                        appName:
                          type: string
                        bundleId:
                          type: string
                        isRequired:
                          type: boolean
                        assignedAt:
                          type: string
                          format: date-time
        '404':
          $ref: '#/components/responses/NotFound'
...
  /apps:
    get:
      operationId: listApps
      summary: List apps
      tags: [apps]
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Paginated list of apps
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppListResponse'
    post:
      operationId: createApp
      summary: Create an app
      tags: [apps]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAppRequest'
      responses:
        '201':
          description: App created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/App'
  /apps/{id}:
    get:
      operationId: getApp
      summary: Get app by ID
      tags: [apps]
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: App details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/App'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      operationId: updateApp
      summary: Update an app
      tags: [apps]
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAppRequest'
      responses:
        '200':
          description: App updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/App'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      operationId: deleteApp
      summary: Delete an app
      tags: [apps]
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '204':
          $ref: '#/components/responses/NoContent'
  /apps/{id}/versions:
    post:
      operationId: createAppVersion
      summary: Add an app version
      tags: [apps]
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAppVersionRequest'
      responses:
        '201':
          description: Version created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppVersion'
        '404':
          $ref: '#/components/responses/NotFound'
  /apps/{id}/assignments:
    post:
      operationId: createAppAssignment
      summary: Assign an app to a device or group
      tags: [apps]
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppAssignmentRequest'
      responses:
        '201':
          description: Assignment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppAssignment'
        '400':
          $ref: '#/components/responses/ValidationError'

  # ==================== KIOSK PROFILES ====================
...